requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
httpx==0.24.1
feedparser>=6.1.0
beautifulsoup4==4.12.2
Pillow>=10.1.0
fastapi>=0.110.0
uvicorn>=0.29.0